% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRead.R
\name{dataRead}
\alias{dataRead}
\title{Read in data and validate before analysis is conducted}
\usage{
dataRead(
  dataFrame,
  timing,
  streams,
  VoI = NA,
  type = "long",
  median0 = NA,
  delta = 3
)
}
\arguments{
\item{dataFrame}{A user inputted dataframe, can be wide or long}

\item{timing}{A string of the timing variable name}

\item{streams}{A string of the streams variable name}

\item{VoI}{A string of the Variable of Interest name}

\item{type}{A string of the type of data (default long)}

\item{median0}{A value for expected median}

\item{delta}{A value for delta (default 3)}
}
\value{
A validated dataframe in long format
}
\description{
Read in data and validate before analysis is conducted
}
\examples{
df <- testData
dataRead(df, timing="hour", streams=c("rep1", "rep2", "rep3", "rep4", "rep5", "rep6", "rep7", "rep8", "rep9", "rep10"), type="wide", median0 = .8)

}
