% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testData}
\alias{testData}
\title{Simulated Call Center Data data}
\format{
\subsection{\code{testData}}{

A data frame with 400 rows and 11 columns:
\describe{
\item{hour}{Hour of the work week}
\item{rep1 - rep10}{Minutes rep spent on phonecall}
\item{year}{Year}
}
}
}
\usage{
testData
}
\description{
A wide format dataset simulated from an example of a call center
for the purpose of testing control chart generation. The 10 reps
are observed during a work week of 40 hours and the duration of
each call is observed in minutes.
}
\keyword{datasets}
